modify restaurant schema

+++ server/services/food/models/restaurant.js
@@ -36,8 +36,8 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     image_url: DataTypes.STRING,
-    longitude: DataTypes.INTEGER,
-    latitude: DataTypes.INTEGER
+    longitude: DataTypes.DECIMAL,
+    latitude: DataTypes.DECIMAL
   }, {
     sequelize,
     modelName: 'Restaurant',
